refactor(tx-flow): move all files from `tx-flow-2` folder to `tx-flow`
