+++ apps/web/src/components/tx-flow/TxFlowStep.tsx
@@ -1,5 +1,5 @@
 import React, { type ReactNode, useContext, useEffect } from 'react'
-import { TxFlowContext, type TxFlowContextType } from './TxFlowProvider'
+import { TxFlowContext, type TxFlowContextType } from '../tx-flow/TxFlowProvider'
 
 export type TxFlowStepProps = TxFlowContextType['txLayoutProps'] & { children?: ReactNode }
