+++ apps/web/src/components/tx-flow/features/TxChecks.tsx
@@ -2,7 +2,7 @@ import type { ReactElement } from 'react'
 import { useContext } from 'react'
 import madProps from '@/utils/mad-props'
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
-import { TxFlowContext } from '@/components/tx-flow-2/TxFlowProvider'
+import { TxFlowContext } from '@/components/tx-flow/TxFlowProvider'
 import useIsCounterfactualSafe from '@/features/counterfactual/hooks/useIsCounterfactualSafe'
 import TxChecksComponent from '@/components/tx/SignOrExecuteForm/TxChecks'
